chore: rename nature snail code file, add docs
